refactor: move legend to types/xy_chart/legend
