+++ src/components/legend/legend.tsx
@@ -2,7 +2,7 @@ import classNames from 'classnames';
 import { inject, observer } from 'mobx-react';
 import React from 'react';
 import { isVertical } from '../../types/xy_chart/utils/axis_utils';
-import { LegendItem as SeriesLegendItem } from '../../lib/series/legend';
+import { LegendItem as SeriesLegendItem } from '../../types/xy_chart/legend/legend';
 import { ChartStore } from '../../state/chart_state';
 import { LegendItem } from './legend_item';
