+++ src/components/react_canvas/bar_geometries.tsx
@@ -2,7 +2,7 @@ import { Group as KonvaGroup } from 'konva';
 import React from 'react';
 import { Group, Rect } from 'react-konva';
 import { animated, Spring } from 'react-spring/renderprops-konva.cjs';
-import { LegendItem } from '../../lib/series/legend';
+import { LegendItem } from '../../types/xy_chart/legend/legend';
 import { BarGeometry, getGeometryStyle } from '../../types/xy_chart/rendering/rendering';
 import { SharedGeometryStyle } from '../../utils/themes/theme';
 import { buildBarRenderProps } from './utils/rendering_props_utils';
