+++ src/types/xy_chart/rendering/rendering.ts
@@ -13,7 +13,7 @@ import { SpecId } from '../../../utils/ids';
 import { isLogarithmicScale } from '../../../utils/scales/scale_continuous';
 import { Scale, ScaleType } from '../../../utils/scales/scales';
 import { CurveType, getCurveFactory } from '../../../utils/curves';
-import { LegendItem } from '../../../lib/series/legend';
+import { LegendItem } from '../legend/legend';
 import { DataSeriesDatum } from '../../../lib/series/series';
 import { belongsToDataSeries } from '../utils/series_utils';
 import { DisplayValueSpec } from '../../../lib/series/specs';
